+++ DefaultSubreddits.h
@@ -1,0 +1,2 @@
+extern NSString *const defaultRandomSubredditsSource;
+extern NSString *const defaultTrendingSubredditsSource;