+++ DefaultSubreddits.m
@@ -1,0 +1,5 @@
+#import <UIKit/UIKit.h>
+#import "DefaultSubreddits.h"
+
+NSString *const defaultRandomSubredditsSource = @"https://jeffreyca.github.io/subreddits/popular.txt";
+NSString *const defaultTrendingSubredditsSource = @"https://jeffreyca.github.io/subreddits/trending-gummy-daily.txt";